feat: Separation of client component from server component #6
